pyproject.toml
[project]
name = "repo-pulse"
version = "0.1.0"
description = "GitHub trending digest and Feishu research assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "apscheduler>=3.11.0",
  "sqlmodel>=0.0.24",
  "httpx>=0.28.0",
  "lark-oapi>=1.4.19",
  "pydantic-settings>=2.10.0",
  "openai>=1.108.0",
  "dashscope",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "pytest-asyncio>=1.1.0",
  "respx>=0.22.0",
]

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
